Rewrite completely.
Well, I've redone it all. Has mostly the same api, but now some
modifiers are only available on the right type of validator. For
example, the 'interval' mod only exists on the number type now.

The new structure makes it easy to add more scoped mods like that,
and allows for me to add new properties to each validator so that
we can log errors or perform mutations on the input.